# docs/swagger.yml
openapi: 3.0.1
info:
  title: Todo API
  description: API for managing todos.
  version: 1.0.0
servers:
  - url: http://localhost:8080/api/v1
paths:
  /todos:
    get:
      summary: Get all todos
      responses:
        '200':
          description: A list of todos
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Todo'
    post:
      summary: Create a new todo
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TodoInput'
      responses:
        '201':
          description: The created todo
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Todo'
  /todos/{id}:
    put:
      summary: Update an existing todo
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TodoInput'
      responses:
        '200':
          description: The updated todo
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Todo'
    delete:
      summary: Delete a todo
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: No content

components:
  schemas:
    Todo:
      type: object
      properties:
        id:
          type: string
          example: "1"
        title:
          type: string
          example: "Buy groceries"
        completed:
          type: boolean
          example: false
    TodoInput:
      type: object
      properties:
        title:
          type: string
        completed:
          type: boolean